Move NBT module into protocol
