+++ src/main.rs
@@ -26,7 +26,6 @@ pub mod macros;
 pub mod ecs;
 pub mod protocol;
 pub mod format;
-pub mod nbt;
 pub mod item;
 pub mod gl;
 pub mod types;
