+++ src/protocol/nbt/mod.rs
@@ -16,8 +16,8 @@ use std::collections::HashMap;
 use std::io;
 use std::io::Read;
 
-use super::protocol::Serializable;
-use super::protocol;
+use crate::protocol::Serializable;
+use crate::protocol;
 use byteorder::{BigEndian, WriteBytesExt, ReadBytesExt};
 
 #[derive(Debug, Clone)]
